zed_extension_api: Fork new version of extension API
