+++ crates/extension_api/Cargo.toml
@@ -1,12 +1,13 @@
 [package]
 name = "zed_extension_api"
-version = "0.3.0"
+version = "0.4.0"
 description = "APIs for creating Zed extensions in Rust"
 repository = "https://github.com/zed-industries/zed"
 documentation = "https://docs.rs/zed_extension_api"
 keywords = ["zed", "extension"]
 edition.workspace = true
-publish = true
+# Change back to `true` when we're ready to publish v0.4.0.
+publish = false
 license = "Apache-2.0"
 
 [lints]
