+++ crates/extension_api/src/extension_api.rs
@@ -198,7 +198,7 @@ mod wit {
 
     wit_bindgen::generate!({
         skip: ["init-extension"],
-        path: "./wit/since_v0.3.0",
+        path: "./wit/since_v0.4.0",
     });
 }
